ConfirmTrade() bugfix with gil

+++ src/map/lua/lua_baseentity.cpp
@@ -4171,7 +4171,7 @@ inline int32 CLuaBaseEntity::confirmTrade(lua_State* L)
             CItem* PItem = PChar->TradeContainer->getItem(slotID);
             if (PItem)
             {
-                uint8 confirmedItems = PChar->TradeContainer->getConfirmedStatus(slotID);
+                uint32 confirmedItems = PChar->TradeContainer->getConfirmedStatus(slotID);
                 auto quantity = (int32)std::min<uint32>(PChar->TradeContainer->getQuantity(slotID), confirmedItems);
 
                 PItem->setReserve(PItem->getReserve() - quantity);
